refactor(bridge): reorganize BridgeView utility functions and improve test structure
